Init project with Next.js and Tailwindcss. Added Fibonacci algorithms
in JavaScript and Python for a speed comparison between the programming
languages. While a single problem isn't enough to justify a hypothesis,
the sole motivation behind this project is to show that JavaScript is
actually easier to read/write than Python, it is faster, and is the
better choice.

Algorithms currently implemented in JavaScript and Python:
- Fibonacci
    - Iterative
    - Recursive
    - Binet's Formula
    - Recursive Memoized