Basic Plotly Chart comparing fibonacci series algorithms in py and js
has been added to the home page. I have not been able to use plotly's
factory function to create the chart. Not sure if it has something
to do with Next but I keep getting a self is not defined error.

Additionally, because of Next, Plotly has to be dynamically loaded.
Which is proving to be a bit awkward awaiting a promise. Currently all
the data for the comparisons is loaded into one chart, the scale
differences between the recursive and performant algorithms makes it so
that you can't really see all the data at once. This will be refactored
in the future.To provide a better experience.