Refactored existing implementation before going further.
The fibonacci comparison plot requires a lot of data. Initially,
building the data either at the API or in SSR was making the data
massive, over 4MB.

This was corrected by using a useEffect hook on the parent page to
build the data. The x values were removed as they were all the same
this reduced the data sent from 4MB to 163KB. This is still too large
for SSR, static generations, or a single fetch for performance reasons.

The next state of the refactor is to split the data into smaller chunks.

The refactor isn't yet complete but the existing implementation has been
modified to be reusable. Rather than send all the data at once,
the current graph will be split into multiple smaller graphs and will
be SSR with the data needed for a single view. Any subsequent data
needed will be fetched on request but saved in the state.

An API has been created in the utils folder that will provide abstracted
methods for fetching data. These calls will act like a "controller"
in the traditional sense. But will have methods available to granularly
fetch our data.