Working on the dynamic Fibonacci chart. The chart loads the default
view correctly, and changes state when a different view is selected.
However, the Plotly chart isn't updating and I have to dig deeper to
find out why exactly. I placed logs in the Class component and can see
that the component registers it should change and as far as I could
force a change was with the title. The Plotly React docs are garbage.
Maybe if you are familiar with Plotly js they are decent but there is
literally no documentation for the react component and gained more
insight from reading the source code. According to the docs the state
might not always update when providing new data and layout values. Which
is crazy because these values are held in the Class component's state.

The data is set in a parent component so the view should re-render
regardless. I am not sure if this is a bug or a feature.